refactor: some refactor and fix typos
