+++ src/components/Header.tsx
@@ -54,7 +54,7 @@ const Header = ({ menuState, toggleMenu, handleOpen }: IHeaderProps) => {
           className='text-xl bg-orange-400/90 rounded-full px-3 py-1 transition-colors hover:bg-orange-500'
           onClick={() => dispatch(actions.logoutUser())}
         >
-          Logout
+          LogOut
         </button>
       ) : (
         <button
@@ -62,7 +62,7 @@ const Header = ({ menuState, toggleMenu, handleOpen }: IHeaderProps) => {
           className='text-xl bg-orange-400/90 rounded-full px-3 py-1 transition-colors hover:bg-orange-500'
           onClick={handleOpen}
         >
-          Login
+          LogIn
         </button>
       )}
     </div>
